requirements.txt
python-telegram-bot
python-dotenv
aiohttp
redis